refactor(checker): extract conditional-expression typing into its own module
helpers.rs reached the 2000-line checker file limit, so move the
conditional (ternary) expression typing — branch typing, dead-branch
diagnostic suppression, and the subtype-reduction collapse — into a new
conditional_expr.rs module. Behavior-preserving; uses the established
`use ...::common`/`::assignability` import pattern so the query-boundary
common-reference count stays net-neutral.
